backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.55.3
httpx==0.27.2
python-dotenv==1.0.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
pgvector==0.3.6
pytest==8.3.5
pytest-cov==6.1.0
pytest-asyncio==0.24.0
psycopg2-binary==2.9.7
azure-identity==1.17.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pypdf==4.2.0
python-docx==1.1.2
beautifulsoup4==4.12.3
playwright==1.44.0
google-api-python-client==2.132.0
google-auth-oauthlib==1.2.0
spacy>=3.7,<3.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
